+++ Changelog.md
@@ -1,3 +1,9 @@
+
+Version 1.8.8
+  * starting a search without defined in and output files gives a better response
+  * BugFix: correct the behaviour if a modification is defined nterm as part of the specificity
+  * BugFix: mzML reader not assigning titel to spectra
+
 Version 1.8.7
   * skyline converter writes crosslinked ssl file